Fix text alignment besides wrapped figures
The LaTeX \raggedright command (and the corresponding flushleft environment)
has a tendency to miss the “never” part,
and will often create ridiculously short lines,
for some minor benefit later in the paragraph.

The Plain TeX version of the command doesn’t suffer this failing,
but is rather conservative: it is loath to create too large a gap
at the end of the line, but in some circumstances
— such as where hyphenation is suppressed — painfully
large gaps may sometimes be required.

Martin Schröder’s ragged2e package offers the best of both worlds:
it provides raggedness which is built on the Plain TeX model,
but which is easily configurable

Via https://texfaq.org/FAQ-ragright